Add flow class
Flow class and host class work together to generate the list of packets
that the host needs to send

Packets are initialized with a reference to their flow. Host calls
packet.flow.log_packet_received() when it receives a non-ack packet.

Flow collects timestamps for when its data packets are received. This
will allow us to generate a time-trace for data packet rate for each
flow.