Remove randomness from simulation
Previously, randomness was used to determine which direction a link
sends, if the front packets on each of the buffers share the same age.
It seemed like this could lead to bad behavior. Now, if both directions
are tied in this way, the direction that sent least recently is chosen.

Also, actors are no longer shuffled by the event manager. So, there is
no randomness in the simulation.